+++ include/sparrow/arrow_interface/arrow_array/private_data.hpp
@@ -18,8 +18,8 @@
 #include <vector>
 
 #include "sparrow/buffer.hpp"
-#include "sparrow/c_interface/arrow_array/smart_pointers.hpp"
-#include "sparrow/c_interface/arrow_array_schema_utils.hpp"
+#include "sparrow/arrow_interface/arrow_array/smart_pointers.hpp"
+#include "sparrow/arrow_interface/arrow_array_schema_utils.hpp"
 
 namespace sparrow
 {
