+++ test/test_arrow_array.cpp
@@ -16,7 +16,7 @@
 #include <memory>
 #include <ranges>
 
-#include "sparrow/c_interface/arrow_array.hpp"
+#include "sparrow/arrow_interface/arrow_array.hpp"
 
 #include "doctest/doctest.h"
